+++ R/aio.R
@@ -282,11 +282,11 @@ unresolved <- function(aio) .Call(rnng_unresolved, aio)
 
 #' Technical Utility: Query if an Aio is Unresolved
 #'
-#' Query whether an Aio remains unresolved. This function is a specialist
+#' Query whether an Aio remains unresolved. This function is a technical utility
 #'     version of \code{\link{unresolved}} not intended for ordinary use.
-#'     Provides a fast method of querying the busy status of an Aio and does not
-#'     alter its state in any way i.e. does not attempt to retrieve any result
-#'     status or messages.
+#'     Provides a method of querying the busy status of an Aio without altering
+#'     its state in any way i.e. not attempting to retrieve any results status
+#'     or messages.
 #'
 #' @param aio an Aio (object of class 'sendAio' or 'recvAio').
 #'